requirements.txt
amqp==2.2.2
Babel==2.5.1
billiard==3.5.0.3
cachetools==2.0.1
celery==4.1.0
certifi==2017.11.5
chardet==3.0.4
configparser==3.5.0
dill==0.2.7.1
Django==2.0
djangorestframework==3.7.4
flower==0.9.2
future==0.16.0
gapic-google-cloud-datastore-v1==0.15.3
gapic-google-cloud-error-reporting-v1beta1==0.15.3
gapic-google-cloud-logging-v2==0.91.3
google-api-core==0.1.3
google-api-python-client==1.6.4
google-auth==1.2.1
google-cloud==0.32.0
google-cloud-bigquery==0.28.0
google-cloud-bigquery-datatransfer==0.1.0
google-cloud-bigtable==0.28.1
google-cloud-container==0.1.0
google-cloud-core==0.28.0
google-cloud-datastore==1.4.0
google-cloud-dns==0.28.0
google-cloud-error-reporting==0.28.0
google-cloud-firestore==0.28.0
google-cloud-language==1.0.0
google-cloud-logging==1.4.0
google-cloud-monitoring==0.28.0
google-cloud-pubsub==0.30.1
google-cloud-resource-manager==0.28.0
google-cloud-runtimeconfig==0.28.0
google-cloud-spanner==0.29.0
google-cloud-speech==0.30.0
google-cloud-storage==1.6.0
google-cloud-trace==0.17.0
google-cloud-translate==1.3.0
google-cloud-videointelligence==1.0.0
google-cloud-vision==0.29.0
google-gax==0.15.16
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpc-google-iam-v1==0.11.4
grpcio==1.8.2
httplib2==0.10.3
idna==2.6
kombu==4.1.0
oauth2client==3.0.0
ply==3.8
proto-google-cloud-datastore-v1==0.90.4
proto-google-cloud-error-reporting-v1beta1==0.15.3
proto-google-cloud-logging-v2==0.91.3
protobuf==3.5.0.post1
psutil==5.4.2
psycopg2==2.7.3.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pytz==2017.3
redis==2.10.6
requests==2.18.4
rsa==3.4.2
six==1.11.0
tornado==4.5.2
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
